Tidy up compile warnings

+++ Src/DasherCore/DasherModel.cpp
@@ -74,7 +74,7 @@ void CDasherModel::Reparent_root(int lower, int upper)
     return;
 
   /* Determine how zoomed in we are */
-  float scalefactor=(m_Rootmax-m_Rootmin)/(upper-lower);
+  float scalefactor=float((m_Rootmax-m_Rootmin)/(upper-lower));
   
   m_Rootmax=int(m_Rootmax+((1024-upper)*scalefactor));
   m_Rootmin=int(m_Rootmin-(lower*scalefactor));
